Fix the code to pass the test
Two changes:

Each thread now keeps its own Lamport clock, syncing with the master
clock when the thread joins or when threads locally communicate.

When large messages are automatically split into chunks in order to get
the start of them onto the wire while the rest are still being computed
(but no messages are being received), all the chunks should have the
Lamport clock that was current when the message started to be created.