Add a timing test for DPF evalution
Useful for comparing time to expand a DPF vs time to load it from disk.

At least on this dev machine, loading it from disk is much faster.

Also add better option parsing, and add a "-c" preprocessing option
which says to store the DPFs compressed (the default is now to store
them expanded).