Add explicit-only mode
Only using [] with explicit (address_t) indices are allowed in this
mode.  Using [] with RegAS or RegXS indices will automatically turn off
this mode, or you can turn it off explicitly.  In explicit-only mode,
updates to the memory in the Shape will not induce communication to the
server or peer, but when it turns off, a message of the size of the
entire Shape will be sent to each of the server and the peer.  This is
useful if you're going to be doing multiple explicit writes to every
element of the Shape before you do your next oblivious read or write.
Bitonic sort is a prime example.