Use only a single (wide) DPF for related reads and writes
It is common that one might make multiple accesses to the ORAM with the
same or related indices, such that everyone knows the _relation_, but
not the absolute indices.  For example, the algorithm may be accessing
a random node in a binary tree (laid our linearly in an array), but also
its two children, whose indices are easily computed from the index of
the parent.

In this case, we can get away with using just a single DPF of width W.
This one DPF can be used for arbitrarily many reads of related indices
(where the relation is revealed, but not the indices), and for up to W
updates.

The current code limits W to at most 5, but this is straightforward to
change.

Also, the current code, although it _uses_ only one DPF, still
_evaluates_ that DPF multiple times (once for each read or update).
This is possible to address, but will save for later.